feat: add honeypot to form
This PR includes adding an input field and label which act as a
honeypot. The name weburl is used as the name of the input field,
as bots tend to recognise names such as 'bot' and other bot-related
terms. Autocomplete is set to off. Used a div to conceal the fields,
instead of type=hidden as bots can often detect this.